backend/requirements.txt
Django==4.2.3
django-cors-headers==4.2.0
django-ninja==0.22.2
django-ninja-extra==0.19.1
django-ninja-jwt==5.2.5
huey==2.4.5
imap-tools==1.0.0
parsel==1.8.1
python-dateutil==2.8.2
python-dotenv==1.0.0
